Feat(Weapons): overhaul weapon pickups
turned weapon pickup into a class that the instances inherit from. Created new component to hold and control hovering and spinning so we don't have that crammed and copy pasted into each one. Made the code to bounce the models work. Added logic to player class to hold an enum to help with the weapon swapping logic I'm doing later.